presence_app/auth_service/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
werkzeug==3.0.1
